+++ igl_cppPort/igl_functions.h
@@ -37,7 +37,7 @@ void igl_boundary_loop(int* F, int nF, int* adjLst, int& sz);
 RH_C_FUNCTION
 void extractIsoLinePts(float* V, int nV, int* F, int nF,
   int* con_idx, double* con_value, int numCon,
-  int divN, float* isoLnPts, int* numPtsPerLst);
+  int divN, double* isoLnPts, int* numPtsPerLst);
 
 RH_C_FUNCTION
 void computeLaplacian(float* V, int nV, int* F, int nF,
